astlpc: Force FW2AHB to map reserved memory
Traditionally the job of handling this mapping has been the
responsibilty of another daemon, but it's less weird if we just enforce
the requirement here were it matters. Anyone using the FW2AHB has to do
so in a coordinated fashion, and the operation is idempotent.

This allows us to avoid taking an unnatural dependency on other daemons
who may also configure the bridge.

Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Change-Id: Ic4e0ef8956e9b1d32e557e73d58058e461c372df