Set source EID from cmdlineand some cleanups
This could be useful when user want to set the Soruce Endpoint ID
of the BMC from the command line argument.

Introduced macros for the following:
Enable thread to send boot complete and  periodic heartbeat
    - MCTP_SPI_SPB_INTERFACE          1

Enable this only when user want to send via sockets
    - MCTP_SPI_USR_SOCKET_ENABLE      1

Delay for Heartbeat signal
    - MCTP_SPI_HEARTBEAT_DELAY        10

Also, rearranged the keepalive function to the right file so that
the corresponding variables could be consumed.