MCTP SPI control daemon initial version
This application could run as a standalone commandline mode
as well as in daemon mode.

In commandline mode, the user could send/receive raw SPI data as
well as some of the Glacier specific (SPB) commands.

Whereas in daemon mode, it is mainly used for sending periodic
Heartbeat signal to the end device (Glacier)

Intruduced GPIO Thread for Interrupt monitoring and enabled
Interrupt function in SPB library

More details are available in ctrld-spi/README file