Dynamic sleep time based on EID pool size
The MCTP control daemon need to sleep based on the Endpoint ID
pool size.

This is recommended from the FPGA team to make sure the delay
provides enough time to generate the routing tables.

Also, change the bootup delay type to integer so that it can handle
more than 256 seconds.