Fix SPI Enumeration
This commit fixes two regressions that happened because of the unification of
the control daemons and SMBus support.

* Fix the SPI message offset macro. The wrong definition was causing us to send
an extra byte leading to timeouts.

* In the error path, the SPI discovery will try to create error logs, but that
requires the dbus object to be initialized. Hence, move the dbus initialization
prior to the discovery process.

Fixes nvbug https://nvbugs/4154648

Tested:
Tested that SPI discovery now works and we get the SPI endpoint D-Bus objects
populated as expected.

Signed-off-by: Santosh Puranik <spuranik@nvidia.com>