ctrld: Add Duplicate EID Removal Logic
This commit is a partial application of a G+H platform patch that
removes duplicate EIDs from the FPGA returned routing table.

Having duplicate EIDs in the routing table causes D-Bus object creation
to fail and the control daemon to exit (as it tries to create an already
existing object again).

This behavior is controlled using a command line switch (-c) to the
control daemon, so that it can only be selectively enabled in the
platforms that need it.

When merging this, we'd have to remove the patch file from G+H
platforms.

Tested: Made sure ctrld works as before.

Change-Id: I3fe4f9d9abb9c7a7252583bf5970afccaa5a21be
Signed-off-by: Santosh Puranik <spuranik@nvidia.com>