demux: Fix message type comparison.
The first message byte in the MCTP payload is the message type and
integrity check (IC) byte.

The most-significant bit is the IC bit and must be masked off when the
demux daemon compares the type to those of registered client types.

This used to work so far because PLDM, SPDM and VDMs don't set the IC
bit. NVMe-MI, however, requires the IC bit to be set.

This commit masks the IC bit off before determining which client to
forward the received message to.

Fixes nvbug https://

Signed-off-by: Santosh Puranik <spuranik@nvidia.com>