demux: Add watchdog
Add a service level watchdog for the demux daemons.
If the service does not notify the watchdog at least once every 10s,
systemd will restart the service.

This is intended to cover cases where the service may be stuck for
extended periods.

Tested:

Instrumented the service code to block it for more than 10s and ensured
watchdog fires.

Performed multiple code updates to ensure no impact on code update
times.

Change-Id: Iba41f03fe2e91e957638e540b755f3a717723d88
Signed-off-by: Santosh Puranik <spuranik@nvidia.com>