libmctp: smbus: Fix transport lockout bug
In the SMBus read function, we re-enable the transport so that the next
set of packets can be transmitted after we are done receiving the
response for the last one.

However, in error cases (such as bad payload), the re-enablement would
be skipped.

This commit fixes that problem by removing the disable/enable transport
portion because that was added as an experiment to make type 5 and type
2 work in parallel. We will revist that when we do type 5 debugging in
detail.

It also prints the buffer received in
case of an error where the SMBus header's length does not match the
actual bytes we read.

Fixes nvbug https://

Signed-off-by: Santosh Puranik <spuranik@nvidia.com>