libmctp: smbus: Tweak retry timings.
The MCTP over SMBus binding specification talks about implementing a
"fair idle" algorithm on top of the standard I2C multi-master
arbitration.

Implementing that algorithm in user space is difficult because of the
strict timing requirements and detection os bus collisions needed.

Instead, we have chosen to just adopt a retry after a short sleep
approach to handling errors on the I2C bus.

This commit increases the number of retries for error scenarios to a
1000 with a 20 microsecond sleep between retries for a worst case
scenario of 20ms wait.

This improves the performance of certain use-cases where we encounter a
busy bus, such as running PLDM type 5 FW update and type 2 telemetry at
the same time.

Fixes jira https://jirasw.nvidia.com/browse/DGXOPENBMC-9302

Tested:
Tested in conjunction with a linux DTS MR that adds multi-master line to
applicable I2C buses. Verified that the FW update performance is much
better with this approach.

Signed-off-by: Santosh Puranik <spuranik@nvidia.com>