ctrld: Add option to pass UUID on CLI
The MCTP SPI control daemon assigns a hardcoded UUID to the EID 0, which
is the HMC ERoT if one was not already assigned to it by the FPGA.

This UUID is not actually set to the endpoint, but rather just hosted as
a D-Bus property. This does not work when the UUID that the control
daemon has hardcoded is different that the one the FPGA assigns.

This causes issues when PLDM is updating FW where it ends up matching
two EIDs 0 over the SPI bus and another over the PCIe bus that are, in
reality, just the same device.

Thus, there is a need to make this UUID for EID 0 configurable. This
commit adds a new command line argument `-u` or `--uuid` where the
platform can specify (via an environment file) the UUID that it wants to
assign to EID 0.

Fixes nvbug https://nvbugs/4358696

Tested:
```
Tested this on a Ranger system by passing in the UUID option to the
mctp-ctrl command line:

mctp-ctrl -e 0 -i 3 -t 6 -m 1 -d 20 -v 1 -u "5f7f6a60-3f28-4cc0-be5d-4985d4c52a0c"

Tested and made sure that the EID 0 D-Bus object reports this supplied
UUID:

busctl get-property xyz.openbmc_project.MCTP.Control.SPI /xyz/openbmc_project/mctp/0/0 xyz.openbmc_project.Common.UUID UUID
s "5f7f6a60-3f28-4cc0-be5d-4985d4c52a0c"

```
Change-Id: I32ccad62fb9c8dc9d728310d11f0161d4d1762bd
Signed-off-by: Santosh Puranik <spuranik@nvidia.com>