libmctp: usb: Get USB endpoint addresses from descriptors
The USB specification allows the device to use any number of its choice
in the lower nibble of its endpoint addresses.

We therefore, cannot hardcode the IN and OUT endpoint addressesin the
binding code.

This commit adds code that queries the addresses from the endpoint
descriptor exposed under the MCTP interface of the device.

Fixes JIRA https://jirasw.nvidia.com/browse/DGXOPENBMC-9471

Change-Id: I71623064087ab0334767c69d64525f9491c20c91
Signed-off-by: Santosh Puranik <spuranik@nvidia.com>