demux: Add client support for tag owner bit.
In the client socket API exposed by the demux daemon, the first byte now
represents the message tag and tag owner bits. Bit 3 is the tag owner
and bits 2:0 are the message tag.

Message tag is not used as of now, but this API now allows the clients
to control the value of the tag owner field in the MCTP packets.

Similarly, in the response message to the client, the first byte now
represents the tag owner and the message tag (bits 3:0) as received from
the remote endpoint.

While this commit handles internal users such as MCTP VDM util and MCTP
Control, this is an API breaking change to other clients of the MCTP
demux daemon and should be merged along with the corresponding changes
as such.

Fixes nvbug https://nvbugs/4490616

Tested: Tested mctp-ctrl, mctp-vdm-util, pldm, spdm and nsm commands
with the respective MRs from those repos.

Signed-off-by: Santosh Puranik <spuranik@nvidia.com>