ctrld: Enable AF_MCTP Sockets
Add a new meson option `mctp-in-kernel-enable` to enable AF_MCTP
in-kernel MCTP sockets. This is used as a POC that the existing control
daemon can work with the in-kernel MCTP implementation.

Fixes JIRA https://jirasw.nvidia.com/browse/DGXOPENBMC-12727

Tested:
This was tested on a GB200NVL with the FPGA with in-kernel MCTP stack
enabled and we used a TUN device to route packets to our userspace MCTP
binding.

We verified that the MCTP control service was able to properly enumerate
the FPGA and all downstream MCTP devices.

Signed-off-by: wthai@nvidia.com
Signed-off-by: Santosh Puranik <spuranik@nvidia.com>