ctrld: Handle Discovery Notifies During Discovery Flow
There is a possibility that we receive discovery notify events when we
are in the middle of the discovery flow. This commit simply discards
such events, continuing with the discovery flow and queues the
processing of the discovery notify, aka, re-discovery by arming the
discovery timer.

Fixes NVBUG https://

Tested: This was tested on a GB200NVL and HGXB by forcing discovery
notifies when we were processing one already or doing a re-discovery. In
both cases, the discovery notify was queued up and processed at a later
point.

Signed-off-by: Santosh Puranik <spuranik@nvidia.com>