smbus: Ignore failures in opening outfd
In a muxed topology, we will have multiple /dev/i2c-x paths to open. It
is possible that certain muxed busses may not be available on a system
(for example, if the downstream card is not plugged in).

In such cases, we should allow the SMBus binding init to succeed with
whatever out FDs we can open.

Fixes NVBUG https://nvbugspro.nvidia.com/bug/4719467

Signed-off-by: Santosh Puranik <spuranik@nvidia.com>