ctrld: spi: Send Restart Notify Before Exit
The restart notify VDM is sent to the ERoT upon service exit today by
using a `ExecStop` in the systemd service file. That invokes the
`mctp-vdm-util` command line, which in turn looks for the ObjectMapper
to get the control daemon's service name. This sometimes can cause races
in the shutdown path because the mapper may be stopped before we run the
VDM command.

Fix this by instead invoking the restart notify within the control
daemon, thus alleviating the need for making a mapper call.

Fixes NVBUG https://nvbugspro.nvidia.com/bug/4774566.

Tested:

Tested this on an HGXB setup and verified that FW updates now take
effect after an HMC reboot and that we send out the restart notify
before we exit the MCTP control daemon.

Signed-off-by: Santosh Puranik <spuranik@nvidia.com>