ctrld: in-kernel: Set Socket Timeout
Set a receive timeout on the AF_MCTP socket used in the control
daemon/vdm-util command line.

This ensures that a recv on an unresponsive endpoint does not block us
forever.

Fixes JIRA https://jirasw.nvidia.com/browse/DGXOPENBMC-12570

Tested:
Tested on a GB200NVL-HMC with an unresponsive HMC ERoT. Ensured that the
discovery process, which earlier used to hang is not passing with a
proper timeout on the HMC ERoT EID.

Signed-off-by: Santosh Puranik <spuranik@nvidia.com>