smbus: treat SIGTERM as an EXIT_SUCCESS event
Systemd sends to SIGTERM (15) to terminate PIDs when their services
are no longer needed. To ensure that service state can be "inactive"
but not "failed", we need to treat SIGTERM as a normal action, which
needs to return EXIT_SUCCESS rather than EXIT_FAILURE to systemd.

Fixes nvbug https://nvbugspro.nvidia.com/bug/5056688
Signed-off-by: Chester Lin <chesterl@nvidia.com>