libmctp: usb: Fix hotplug device matching
The hotplug handler for device matching was recently changed to remove
the checks for vendor ID and product ID. This was done as we now need to
support MCUs that can have varying product IDs.

However, the callback now is too broad and does not properly check for
whether the device is an MCTP capable device. This causes issues where a
root hub port is being treated as an MCTP device and hence demux daemon
uses the wrong FD to try to Tx/Rx. This was not seen on our testing
since the order of the device hotplug callback can be non-deterministic.

Fix the matching to ensure we only handle a device that has an MCTP
interface class.

Fixes JIRA https://

Tested: Tested on a GB200NVL HMC setup where the issue was consistently
being seen and we no longer process devices that do not have the MCTP
interface descriptor.

Signed-off-by: Santosh Puranik <spuranik@nvidia.com>