usb: Register hotplug callback after initialization is completed
In the past, the demux is launched before device event arrived.
So the callback will not be executed immediately.
Now the launching is trigger via udev, it mean the callback can be
executed immediately.

So register the callback function at the end of the mctp_usb_init.

Fixes jira https://nvbugspro.nvidia.com/bug/5087706
Signed-off-by: Iris Chen <irchen@nvidia.com>