libmctp: usb: Fix MCTP USB binding header size
The calculation for the binding header size was taking the sizeof the
binding header structure's pointer rather than the structure itself.

This worked on 32-bit machines simply because the size of the binding
header structure is the same as that of a pointer. On 64-bit machines we
end up sending 4 extra bytes instead.

Fix this by taking the sizeof the structure.

Fixes JIRA https://

Tested: Tested and verified on AST2700 EVB that we no longer send
additional bytes in USB Tx.

Signed-off-by: Santosh Puranik <spuranik@nvidia.com>